// src/api/bindings/Settings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BackupSettings } from "./BackupSettings";
import type { BillingSettings } from "./BillingSettings";
import type { GymSettings } from "./GymSettings";
import type { MembershipSettings } from "./MembershipSettings";
import type { SecuritySettings } from "./SecuritySettings";
import type { SystemState } from "./SystemState";
import type { WhatsappSettings } from "./WhatsappSettings";

export type Settings = { gym: GymSettings, membership: MembershipSettings, billing: BillingSettings, whatsapp: WhatsappSettings, backup: BackupSettings, security: SecuritySettings, system: SystemState, };